em_backend/requirements.txt
asgiref==3.8.1
Django
django-cors-headers==4.3.1
django-guardian==2.4.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
PyJWT==1.7.1
PyMySQL==1.1.1
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
uritemplate==4.1.1
python-dotenv
